Added dependencies for RedHat
Added the required build dependencies for building Principia on Redhat and Redhat based distros, such as Fedora. I figured since I am using a Redhat-based distro, I thought I would help since no one else spent time like me figuring out the right dependencies for it to build on Fedora. This has been tested on Fedora 37, and it was able to compile and run. However, it may not work for all Redhat based distros, as I tested this on a Redhat Enterprise virtual machine and some of the packages were not found.