Added 'glew-devel' for Fedora
When I was recompiling Principia on Fedora, I found there was an error when compiling. When I looked on the dependencies for other Linux systems and the compile error message, I found that I was missing 'libglew'. I looked for the package, installed it, and it was able to compile on Fedora.